Ast: rework structure of ast crate
