+++ crates/ast/src/lib.rs
@@ -1,353 +1,12 @@
-use itertools::Itertools;
-use syntax::{SyntaxElement, SyntaxKind, SyntaxNode, SyntaxToken};
+use syntax::{SyntaxKind, SyntaxNode};
 
-#[derive(Debug)]
-pub struct VariableDef(SyntaxNode);
+mod expr;
+mod stmt;
 
-impl VariableDef {
-    pub fn name(&self) -> Option<SyntaxToken> {
-        self.0
-            .children_with_tokens()
-            .filter_map(SyntaxElement::into_token)
-            .find(|token| token.kind() == SyntaxKind::Ident)
-    }
+pub use expr::*;
+pub use stmt::*;
 
-    pub fn value(&self) -> Option<Expr> {
-        self.0.children().find_map(Expr::cast)
-    }
-}
-
-#[derive(Debug)]
-pub struct FunctionDef(SyntaxNode);
-
-impl FunctionDef {
-    pub fn name(&self) -> Option<SyntaxToken> {
-        self.0
-            .children_with_tokens()
-            .filter_map(SyntaxElement::into_token)
-            .find(|token| token.kind() == SyntaxKind::Ident)
-    }
-
-    pub fn param_ident_list(&self) -> Vec<SyntaxToken> {
-        self.0
-            .children()
-            .find(|node| node.kind() == SyntaxKind::ParamIdentList)
-            .map_or(vec![], |n| {
-                n.children_with_tokens()
-                    .filter_map(SyntaxElement::into_token)
-                    .filter(|token| token.kind() == SyntaxKind::Ident)
-                    .collect()
-            })
-    }
-
-    pub fn body(&self) -> Option<CodeBlock> {
-        self.0
-            .children()
-            .find(|t| t.kind() == SyntaxKind::CodeBlock)
-            .map(CodeBlock)
-    }
-}
-
-#[derive(Debug)]
-pub struct FunctionCall(SyntaxNode);
-
-impl FunctionCall {
-    pub fn name(&self) -> Option<SyntaxToken> {
-        self.0
-            .children_with_tokens()
-            .filter_map(SyntaxElement::into_token)
-            .find(|token| token.kind() == SyntaxKind::Ident)
-    }
-
-    pub fn param_value_list(&self) -> Vec<Expr> {
-        self.0
-            .children()
-            .find(|node| node.kind() == SyntaxKind::ParamValueList)
-            .map_or(vec![], |n| n.children().filter_map(Expr::cast).collect())
-    }
-}
-
-#[derive(Debug)]
-pub struct BinaryExpr(SyntaxNode);
-
-impl BinaryExpr {
-    pub fn lhs(&self) -> Option<Expr> {
-        self.0.children().find_map(Expr::cast)
-    }
-
-    pub fn rhs(&self) -> Option<Expr> {
-        self.0.children().filter_map(Expr::cast).nth(1)
-    }
-
-    pub fn op(&self) -> Option<SyntaxToken> {
-        self.0
-            .children_with_tokens()
-            .filter_map(SyntaxElement::into_token)
-            .find(|token| {
-                matches!(
-                    token.kind(),
-                    SyntaxKind::Plus
-                        | SyntaxKind::Minus
-                        | SyntaxKind::Star
-                        | SyntaxKind::Slash
-                        | SyntaxKind::GreaterThan
-                        | SyntaxKind::GreaterThanOrEqual
-                        | SyntaxKind::LessThan
-                        | SyntaxKind::LessThanOrEqual
-                        | SyntaxKind::AndKw
-                        | SyntaxKind::OrKw
-                        | SyntaxKind::ComparisonEquals
-                )
-            })
-    }
-}
-
-#[derive(Debug)]
-pub enum Literal {
-    Number(Number),
-    String(crate::String),
-    Boolean(Boolean),
-}
-
-impl Literal {
-    pub fn cast(node: SyntaxNode) -> Self {
-        match node
-            .first_token()
-            .expect("Literal's must have a child")
-            .kind()
-        {
-            SyntaxKind::Number => Self::Number(Number(node)),
-            SyntaxKind::String => Self::String(String(node)),
-            SyntaxKind::Boolean => Self::Boolean(Boolean(node)),
-            _ => unreachable!("Literals must be Numbers or Strings"),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct Number(SyntaxNode);
-
-impl Number {
-    pub fn parse(&self) -> f64 {
-        self.0
-            .children_with_tokens()
-            .filter(|t| t.kind() == SyntaxKind::Number)
-            .map(|t| t.to_string())
-            .join(".")
-            .parse()
-            .unwrap()
-    }
-}
-
-#[derive(Debug)]
-pub struct String(SyntaxNode);
-
-impl String {
-    pub fn parse(&self) -> std::string::String {
-        self.0.first_token().unwrap().text().to_string()
-    }
-}
-
-#[derive(Debug)]
-pub struct Boolean(SyntaxNode);
-
-impl Boolean {
-    pub fn parse(&self) -> bool {
-        self.0.first_token().unwrap().text().parse().unwrap()
-    }
-}
-
-#[derive(Debug)]
-pub struct ParenExpr(SyntaxNode);
-
-impl ParenExpr {
-    pub fn expr(&self) -> Option<Expr> {
-        self.0.children().find_map(Expr::cast)
-    }
-}
-
-#[derive(Debug)]
-pub struct UnaryExpr(SyntaxNode);
-
-impl UnaryExpr {
-    pub fn expr(&self) -> Option<Expr> {
-        self.0.children().find_map(Expr::cast)
-    }
-
-    pub fn op(&self) -> Option<SyntaxToken> {
-        self.0
-            .children_with_tokens()
-            .filter_map(SyntaxElement::into_token)
-            .find(|token| matches!(token.kind(), SyntaxKind::Minus | SyntaxKind::NotKw))
-    }
-}
-
-#[derive(Debug)]
-pub struct VariableRef(SyntaxNode);
-
-impl VariableRef {
-    pub fn name(&self) -> Option<SyntaxToken> {
-        self.0.first_token()
-    }
-}
-
-#[derive(Debug)]
-pub struct VariableAssign(SyntaxNode);
-
-impl VariableAssign {
-    pub fn name(&self) -> Option<SyntaxToken> {
-        self.0.first_token()
-    }
-
-    pub fn value(&self) -> Option<Expr> {
-        self.0.children().find_map(Expr::cast)
-    }
-}
-
-#[derive(Debug)]
-pub enum Expr {
-    BinaryExpr(BinaryExpr),
-    Literal(Literal),
-    ParenExpr(ParenExpr),
-    UnaryExpr(UnaryExpr),
-    VariableRef(VariableRef),
-    CodeBlock(CodeBlock),
-    FunctionCall(FunctionCall),
-}
-
-impl Expr {
-    pub fn cast(node: SyntaxNode) -> Option<Self> {
-        let result = match node.kind() {
-            SyntaxKind::InfixExpr => Self::BinaryExpr(BinaryExpr(node)),
-            SyntaxKind::Literal => Self::Literal(Literal::cast(node)),
-            SyntaxKind::ParenExpr => Self::ParenExpr(ParenExpr(node)),
-            SyntaxKind::PrefixExpr => Self::UnaryExpr(UnaryExpr(node)),
-            SyntaxKind::VariableRef => Self::VariableRef(VariableRef(node)),
-            SyntaxKind::CodeBlock => Self::CodeBlock(CodeBlock(node)),
-            SyntaxKind::FunctionCall => Self::FunctionCall(FunctionCall(node)),
-            _ => {
-                return None;
-            }
-        };
-
-        Some(result)
-    }
-}
-
-#[derive(Debug)]
-pub struct CodeBlock(SyntaxNode);
-
-impl CodeBlock {
-    pub fn stmts(&self) -> Vec<Stmt> {
-        self.0.children().filter_map(Stmt::cast).collect()
-    }
-}
-
-#[derive(Debug)]
-pub struct If(SyntaxNode);
-
-impl If {
-    pub fn expr(&self) -> Option<Expr> {
-        self.0.children().find_map(Expr::cast)
-    }
-
-    pub fn body(&self) -> Option<CodeBlock> {
-        self.0
-            .children()
-            .find(|t| t.kind() == SyntaxKind::CodeBlock)
-            .map(CodeBlock)
-    }
-
-    pub fn else_branch(&self) -> Option<ElseBranch> {
-        let else_node = self
-            .0
-            .children()
-            .find(|t| t.kind() == SyntaxKind::Else)
-            .map(Else);
-
-        let else_if_node = self
-            .0
-            .children()
-            .find(|t| t.kind() == SyntaxKind::ElseIf)
-            .map(ElseIf);
-
-        if let Some(else_stmt) = else_node {
-            Some(ElseBranch::Else(else_stmt))
-        } else {
-            else_if_node.map(ElseBranch::ElseIf)
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct ElseIf(SyntaxNode);
-
-impl ElseIf {
-    pub fn if_stmt(&self) -> Option<If> {
-        self.0
-            .children()
-            .find(|t| t.kind() == SyntaxKind::If)
-            .map(If)
-    }
-}
-
-#[derive(Debug)]
-pub struct Else(SyntaxNode);
-
-impl Else {
-    pub fn body(&self) -> Option<CodeBlock> {
-        self.0
-            .children()
-            .find(|t| t.kind() == SyntaxKind::CodeBlock)
-            .map(CodeBlock)
-    }
-}
-
-#[derive(Debug)]
-pub enum ElseBranch {
-    Else(Else),
-    ElseIf(ElseIf),
-}
-
-#[derive(Debug)]
-pub struct Return(SyntaxNode);
-
-impl Return {
-    pub fn value(&self) -> Option<Expr> {
-        self.0.children().find_map(Expr::cast)
-    }
-}
-
-#[derive(Debug)]
-pub enum Stmt {
-    VariableDef(VariableDef),
-    Expr(Expr),
-    FunctionDef(FunctionDef),
-    VariableAssign(VariableAssign),
-    If(If),
-    ElseIf(ElseIf),
-    Else(Else),
-    Return(Return),
-}
-
-impl Stmt {
-    pub fn cast(node: SyntaxNode) -> Option<Self> {
-        let result = match node.kind() {
-            SyntaxKind::VariableDef => Self::VariableDef(VariableDef(node)),
-            SyntaxKind::FunctionDef => Self::FunctionDef(FunctionDef(node)),
-            SyntaxKind::VariableAssign => Self::VariableAssign(VariableAssign(node)),
-            SyntaxKind::If => Self::If(If(node)),
-            SyntaxKind::ElseIf => Self::ElseIf(ElseIf(node)),
-            SyntaxKind::Else => Self::Else(Else(node)),
-            SyntaxKind::Return => Self::Return(Return(node)),
-            _ => Self::Expr(Expr::cast(node)?),
-        };
-
-        Some(result)
-    }
-}
-
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct Root(SyntaxNode);
 
 impl Root {
@@ -359,7 +18,7 @@ impl Root {
         }
     }
 
-    pub fn stmts(&self) -> impl Iterator<Item = Stmt> {
-        self.0.children().filter_map(Stmt::cast)
+    pub fn stmts(&self) -> impl Iterator<Item = stmt::Stmt> {
+        self.0.children().filter_map(stmt::Stmt::cast)
     }
 }
